split GraphQL endpoints & schema: /api/ & /docs/ for users, /api/admin/ & /docs/admin/ for the "manage-users" status/role
