+++ src/db/admin/index.js
@@ -1,0 +1,9 @@
+// admin
+const User = require('./user');
+// user
+const Event = require('../event');
+
+module.exports = {
+  User,
+  Event
+}
