+++ src/db/admin/user.js
@@ -1,6 +1,6 @@
 'use strict';
 const Sequelize = require('sequelize');
-const sequelize = require('./connection');
+const sequelize = require('../connection');
 const tools = require('auto-load')('src/tools');
 const bcrypt = require('bcryptjs');
