+++ src/db/index.js
@@ -1,7 +1,5 @@
 const Event = require('./event');
-const User = require('./user');
 
 module.exports = {
-  Event,
-  User
+  Event
 }
