+++ src/mutations/admin/index.js
@@ -1,0 +1,7 @@
+const createEvent = require('../eventCreate');
+const createUser = require('./userCreate');
+
+module.exports = {
+  createEvent,
+  createUser
+}
