+++ src/mutations/index.js
@@ -1,7 +1,5 @@
 const createEvent = require('./eventCreate');
-const createUser = require('./userCreate');
 
 module.exports = {
-  createEvent,
-  createUser
+  createEvent
 }
