+++ src/queries/admin/index.js
@@ -1,0 +1,11 @@
+const Event = require('../event');
+const allEvents = require('../allEvents');
+const User = require('./user');
+const Users = require('./users');
+
+module.exports = {
+  Event,
+  allEvents,
+  User,
+  Users
+}
