+++ src/queries/admin/user.js
@@ -1,6 +1,6 @@
 const { GraphQLID, GraphQLNonNull } = require('graphql')
-const UserType = require('../types/user')
-const getProjection = require('../tools/projection')
+const UserType = require('../../types/user')
+const getProjection = require('../../tools/projection')
 
 module.exports = {
   type: UserType,
