+++ src/queries/index.js
@@ -1,11 +1,7 @@
 const Event = require('./event');
 const allEvents = require('./allEvents');
-const User = require('./user');
-const Users = require('./users');
 
 module.exports = {
   Event,
   allEvents,
-  User,
-  Users
 }
