+++ src/subscriptions/index.js
@@ -1,5 +1,8 @@
-const eventCreated = require('./eventCreated')
+//const userCreated = require('./userCreated');
+const eventCreated = require('./eventCreated');
+// TODO: create admin subscriptions server
 
 module.exports = {
+  //userCreated,
   eventCreated
 }
