+++ src/subscriptions/userCreated.js
@@ -1,0 +1,7 @@
+const socket = require('../socket');
+const UserType = require('../types/user');
+
+module.exports = {
+  type: UserType,
+  subscribe: () => socket.asyncIterator('USER_CREATED')
+}
